seo-dashboard/requirements.txt
# Dependencias para SEO Dashboard API Server
# Instalar con: pip install -r requirements.txt

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# Google APIs
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.110.0

# Utilidades
python-dateutil>=2.8.2
